Move print_memory_usage to cli.rs
